+++ packages/firefly-honeycomb/tests/getInstrumentationOptions.test.ts
@@ -1,6 +1,6 @@
 import type { FetchInstrumentationConfig } from "@opentelemetry/instrumentation-fetch";
 import { FireflyRuntime } from "@squide/firefly";
-import { __clearActiveSpanMock, __setActiveSpanMock } from "../src/activeSpan.ts";
+import { __clearOverrideFetchRequestSpanWithActiveSpanContextMock, __setOverrideFetchRequestSpanWithActiveSpanContextMock } from "../src/activeSpan.ts";
 import { getInstrumentationOptions } from "../src/registerHoneycombInstrumentation.ts";
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -38,7 +38,10 @@ test("when debug is true", () => {
         apiKey: "123"
     });
 
-    const cleanedResult = removeInstrumentationVersionsForSnapshot(result);
+    const cleanedResult = removeInstrumentationVersionsForSnapshot({
+        ...result,
+        fetchInstrumentation: result.fetchInstrumentation ? result.fetchInstrumentation({}) : undefined
+    });
 
     expect(cleanedResult).toMatchSnapshot();
 });
@@ -51,7 +54,10 @@ test("when debug is false", () => {
         apiKey: "123"
     });
 
-    const cleanedResult = removeInstrumentationVersionsForSnapshot(result);
+    const cleanedResult = removeInstrumentationVersionsForSnapshot({
+        ...result,
+        fetchInstrumentation: result.fetchInstrumentation ? result.fetchInstrumentation({}) : undefined
+    });
 
     expect(cleanedResult).toMatchSnapshot();
 });
@@ -65,7 +71,10 @@ test("when the runtime mode is \"development\"", () => {
         apiKey: "123"
     });
 
-    const cleanedResult = removeInstrumentationVersionsForSnapshot(result);
+    const cleanedResult = removeInstrumentationVersionsForSnapshot({
+        ...result,
+        fetchInstrumentation: result.fetchInstrumentation ? result.fetchInstrumentation({}) : undefined
+    });
 
     expect(cleanedResult).toMatchSnapshot();
 });
@@ -79,7 +88,10 @@ test("when the runtime mode is \"production\"", () => {
         apiKey: "123"
     });
 
-    const cleanedResult = removeInstrumentationVersionsForSnapshot(result);
+    const cleanedResult = removeInstrumentationVersionsForSnapshot({
+        ...result,
+        fetchInstrumentation: result.fetchInstrumentation ? result.fetchInstrumentation({}) : undefined
+    });
 
     expect(cleanedResult).toMatchSnapshot();
 });
@@ -113,11 +125,11 @@ describe("fetchInstrumentation", () => {
 
         expect(mock).toHaveBeenCalledTimes(1);
         expect(mock).toHaveBeenCalledWith(expect.objectContaining({
-            applyCustomAttributesOnSpan: expect.any(Function)
+            requestHook: expect.any(Function)
         }));
     });
 
-    test("when fetchInstrumentation is not provided, applyCustomAttributesOnSpan is the active span function", () => {
+    test("when fetchInstrumentation is not provided, requestHook is the active span override function", () => {
         const runtime = new FireflyRuntime();
 
         const result = getInstrumentationOptions(runtime, {
@@ -131,16 +143,16 @@ describe("fetchInstrumentation", () => {
         // @ts-ignore
         const fetchOptions = result.fetchInstrumentation({}) as FetchInstrumentationConfig;
 
-        expect(fetchOptions.applyCustomAttributesOnSpan).toBeDefined();
+        expect(fetchOptions.requestHook).toBeDefined();
     });
 
-    test("when the base honeycomb instrumentation library configure a default applyCustomAttributesOnSpan, merge the base function with the active span function", () => {
+    test("when the base honeycomb instrumentation library configure a default requestHook, merge the base function with the active span override function", () => {
         const runtime = new FireflyRuntime();
 
         const baseConfigMock = jest.fn();
         const activeSpanMock = jest.fn();
 
-        __setActiveSpanMock(activeSpanMock);
+        __setOverrideFetchRequestSpanWithActiveSpanContextMock(activeSpanMock);
 
         const result = getInstrumentationOptions(runtime, {
             apiKey: "123"
@@ -149,15 +161,15 @@ describe("fetchInstrumentation", () => {
         // Simulate calling the "registerHoneycombInstrumentation" function.
         // eslint-disable-next-line @typescript-eslint/ban-ts-comment
         // @ts-ignore
-        const { applyCustomAttributesOnSpan } = result.fetchInstrumentation({
-            applyCustomAttributesOnSpan: baseConfigMock
+        const { requestHook } = result.fetchInstrumentation({
+            requestHook: baseConfigMock
         });
 
-        applyCustomAttributesOnSpan();
+        requestHook();
 
         expect(baseConfigMock).toHaveBeenCalledTimes(1);
         expect(activeSpanMock).toHaveBeenCalledTimes(1);
 
-        __clearActiveSpanMock();
+        __clearOverrideFetchRequestSpanWithActiveSpanContextMock();
     });
 });